chore(src): rename types file
